requirements.txt
requests
datetime
os
python-dotenv